Refactor && address comments
