+++ src/Azure/AzureClientInfo.ts
