+++ src/Azure/WebApp/FunctionAppService.ts
@@ -4,57 +4,54 @@
  *--------------------------------------------------------------------------------------------*/
 import WebSiteManagementClient from "azure-arm-website";
 import { ServiceClientCredentials, WebResource } from "ms-rest";
-import * as request from 'request-promise';
+import { AzureEnvironment } from 'ms-rest-azure';
 import { appendExtensionUserAgent, createAzureClient } from "vscode-azureextensionui";
 import * as Constants from "../../constants";
-import { ApisTreeItem } from "../../explorer/ApisTreeItem";
-import { ApiTreeItem } from "../../explorer/ApiTreeItem";
-import { ext } from "../../extensionVariables";
-import { localize } from "../../localize";
 import { nRequest, requestUtil, sendRequest } from "../../utils/requestUtil";
 import { signRequest } from "../../utils/signRequest";
-import { IAzureClientInfo } from "../Azure/AzureClientInfo";
-import { IFunctionKeys } from "./Function";
+import { IAzureClientInfo } from "../AzureClientInfo";
+import { IFunctionHostKeyResponse, IFunctionKeys } from "./IFunctionContract";
 import { IFunctionContract } from "./IFunctionContract";
 
-export namespace FunctionService {
+export class FunctionAppService {
+    public baseUrl: string;
+
     // Get all functions from a function app
-    export async function listAllFunctions(baseUrl: string, credentials: ServiceClientCredentials): Promise<IFunctionContract[]> {
+    public async listAllFunctions(baseUrl: string, credentials: ServiceClientCredentials): Promise<IFunctionContract[]> {
         const queryUrl = `${baseUrl}/functions?api-version=${Constants.functionAppApiVersion}`;
-        const funcAppObj = await requestUtil(queryUrl, credentials);
+        const funcAppObj : string = await requestUtil(queryUrl, credentials);
         // tslint:disable-next-line: no-unsafe-any
         return JSON.parse(funcAppObj).value;
     }
 
     // check if the imported function app already has the hostkey, otherwise created
-    export async function addFuncHostKey(baseUrl: string, credentials: ServiceClientCredentials, serviceName: string): Promise<string> {
-        const hostKeys = await getFuncHostKeys(baseUrl, credentials);
+    public async addFuncHostKey(baseUrl: string, credentials: ServiceClientCredentials, serviceName: string): Promise<string> {
+        const hostKeys = await this.getFuncHostKeys(baseUrl, credentials);
         const funcAppKeyName = `apim-${serviceName}`;
         if (hostKeys !== undefined && hostKeys.functionKeys !== undefined && hostKeys.functionKeys[funcAppKeyName]) {
             return hostKeys.functionKeys[funcAppKeyName];
         }
-        ext.outputChannel.appendLine(localize("importFunctionApp", `Can't find existing function app host key ${funcAppKeyName}`));
-        return await createFuncHostKey(baseUrl, funcAppKeyName, credentials);
+        return await this.createFuncHostKey(baseUrl, funcAppKeyName, credentials);
     }
 
     // Get function hostkey
-    export async function getFuncHostKeys(baseUrl: string, credentials: ServiceClientCredentials): Promise<IFunctionKeys> {
-        const options = new WebResource();
-        options.method = "POST";
-        options.url = `${baseUrl}/host/default/listkeys?api-version=${Constants.functionAppApiVersion}`;
-        options.headers = {
-            ['User-Agent']: appendExtensionUserAgent()
-        };
-        await signRequest(options, credentials);
-        // tslint:disable-next-line: await-promise
-        const funcKeys = await request(options).promise();
-        // tslint:disable-next-line: no-unsafe-any
-        return JSON.parse(funcKeys);
+    public async getFuncHostKeys(baseUrl: string, credentials: ServiceClientCredentials): Promise<IFunctionKeys> {
+        // const options = new WebResource();
+        // options.method = "POST";
+        // options.url = `${baseUrl}/host/default/listkeys?api-version=${Constants.functionAppApiVersion}`;
+        // options.headers = {
+        //     ['User-Agent']: appendExtensionUserAgent()
+        // };
+        // await signRequest(options, credentials);
+        // // tslint:disable-next-line: await-promise
+        // const funcKeys = await request(options).promise();
+        // // tslint:disable-next-line: no-unsafe-any
+        const url = `${baseUrl}/host/default/listkeys?api-version=${Constants.functionAppApiVersion}`;
+        return await requestUtil(url, credentials, "POST");
     }
 
     // Create function hostkey for our apim service
-    export async function createFuncHostKey(baseUrl: string, funcKeyName: string, credentials: ServiceClientCredentials): Promise<string> {
-        ext.outputChannel.appendLine(localize("importFunctionApp", `Create new function app host key ${funcKeyName}`));
+    public async createFuncHostKey(baseUrl: string, funcKeyName: string, credentials: ServiceClientCredentials): Promise<string> {
         // create new webresource
         const options = new WebResource();
         options.method = "PUT";
@@ -69,7 +66,7 @@ export namespace FunctionService {
             properties: {}
         };
         funKeyRequest.json = true;
-        const response = await sendRequest(funKeyRequest);
+        const response: IFunctionHostKeyResponse = await sendRequest(funKeyRequest);
         if (response.properties.value) {
             return String(response.properties.value);
         } else {
@@ -78,12 +75,17 @@ export namespace FunctionService {
     }
 
     // Create client for webiste mgmt client
-    export function getClient(node: ApisTreeItem | ApiTreeItem): WebSiteManagementClient {
+    public getClient(credentials: ServiceClientCredentials, subscriptionId: string, environment: AzureEnvironment): WebSiteManagementClient {
         const clientInfo: IAzureClientInfo = {
-            credentials: node.root.credentials,
-            subscriptionId: node.root.subscriptionId,
-            environment: node.root.environment
+            credentials: credentials,
+            subscriptionId: subscriptionId,
+            environment: environment
         };
         return createAzureClient(clientInfo, WebSiteManagementClient);
+    }
+
+    // Get site url
+    public getSiteUrl(endPointUrl: string, subscriptionId: string, resourceGroup: string, functionName: string): string {
+        return `${endPointUrl}/subscriptions/${subscriptionId}/resourceGroups/${resourceGroup}/providers/Microsoft.Web/sites/${functionName}`;
     }
 }
