+++ src/extension.ts
@@ -13,8 +13,8 @@ import { copySubscriptionKey } from './commands/copySubscriptionKey';
 import { createService } from './commands/createService';
 import { deleteNode } from './commands/deleteNode';
 import { extractAPI, extractService } from './commands/extract';
-import { importFunctionApp } from './commands/importFunctionApp';
-import { importFunctionAppToApi } from './commands/importFunctionApp';
+import { importFunctionApp } from './commands/importFunctionApp/importFunctionApp';
+import { importFunctionAppToApi } from './commands/importFunctionApp/importFunctionApp';
 import { importOpenApi } from './commands/importOpenApi';
 import { createNamedValue, updateNamedValue } from './commands/manageNamedValue';
 import { openInPortal } from './commands/openInPortal';
